Refactor previewers
